Fixed issue when trying to add item in mobile view (full-screen dialog)
Adding an item on the generic multifield doesn't work if the dialog is
opened in full-screen within the mobile view. The draw-back was that the
item is created before opening the dialog but never got deleted if it
fails.

The fix is to delete the item again if an error occurs during opening the
dialog.